fixed interface for Alter command
